app/python2.7/requirements.txt
websocket_client==0.57.0

-e git://github.com/nsg-ethz/p4-utils.git#egg=p4utils
